web/requirements.txt
argon2-cffi==25.1.0
Flask==3.1.1
psycopg2==2.9.10
python-dotenv==1.1.0
pytest==8.4.0